# Support pruning subgraphs from ancestor traversal.