This is meant to allow an upcoming `commit-graph write` command use the
existing traversal code to find new commits that haven't already been
indexed in an existing commit graph file.

It may also be useful for incremental pack generation.

I'm not exactly happy with the whole `OidPredicate` business, but I couldn't think of a better way.