@avoidscorn 2021-05-23
I replaced the `OidPredicate` stuff with just using `fn(&oid) -> bool` as the type parameter for the default include-all-ancestors case.

Abridged `traversal` numbers (disabled tree traversal and libgit2 steps) show no noticeable changes in performance on my system:
* `main`:
  ```
  $ command time -v cargo run --release --package traversal -- ~/src/linux/.git master
  [...]
  gitoxide (uncached): collect all 1013596 commits in 9.306947981s (108907 commits/s)
  gitoxide (cache = 57MB): confirmed 1013596 commits in 9.255322714s (109515 commits/s)
  gitoxide (static cache = 64 entries): confirmed 1013596 commits in 9.370033327s (108174 commits/s)
  gitoxide (uncached): confirmed 1013596 commits in 9.254398583s (109526 commits/s)
  [...]
	Maximum resident set size (kbytes): 846692
  ```
* With `OidPredicate`:
  ```
  $ command time -v cargo run --release --package traversal -- ~/src/linux/.git master
  [...]
  gitoxide (uncached): collect all 1013596 commits in 9.268462226s (109360 commits/s)
  gitoxide (cache = 57MB): confirmed 1013596 commits in 9.25482569s (109521 commits/s)
  gitoxide (static cache = 64 entries): confirmed 1013596 commits in 9.451119544s (107246 commits/s)
  gitoxide (uncached): confirmed 1013596 commits in 9.307447228s (108902 commits/s)
  [...]
	Maximum resident set size (kbytes): 846784
  ```
* With `fn(&oid) -> bool`:
  ```
  $ command time -v cargo run --release --package traversal -- ~/src/linux/.git master
  [...]
  gitoxide (uncached): collect all 1013596 commits in 9.268462226s (109360 commits/s)
  gitoxide (cache = 57MB): confirmed 1013596 commits in 9.25482569s (109521 commits/s)
  gitoxide (static cache = 64 entries): confirmed 1013596 commits in 9.451119544s (107246 commits/s)
  gitoxide (uncached): confirmed 1013596 commits in 9.307447228s (108902 commits/s)
  [...]
	Maximum resident set size (kbytes): 846784
  ```