@avoidscorn 2021-05-25
I'd just as soon keep using PRs for all of my changes. But if that is a burden for you, I'm open to direct pushes for `git-commitgraph`-only changes.