@Byron 2021-05-25
Great, I invited you! Welcome to `gitoxide` 🎉🙌! 

Please feel free to do PRs for `git-commitgraph` related changes if you want feedback - it's just that I do trust you delivering great software on your own and want to show that by inviting you to the collaborator group.

I myself will also do only minor changes to `git-commitgraph` without consulting you, but for anything akin to API changes or feature additions, I would start a discussion and/or create PRs to get feedback early. Everything else we can certainly flesh out as we go :D.